Restructuration du module categories
